# Optimize console poseidon sponge apply_s_box