For 2000 transactions
ViewKey 1 is_owner - took 1016ms
ViewKey 2 is_owner - took 983ms
ViewKey 3 is_owner - took 979ms
ViewKey 4 is_owner - took 995ms

For 2000 transactions
ViewKey 1 is_owner - took 981ms
ViewKey 2 is_owner - took 966ms
ViewKey 3 is_owner - took 960ms
ViewKey 4 is_owner - took 974ms

`- 2.3%`

![image](https://user-images.githubusercontent.com/34972409/197878411-7449835e-dd6c-46aa-b7dc-5e39880cfb2a.png)